requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
pytest>=7.4.0
